requirements.txt
streamlit==0.82.0
requests
PyYAML==5.4.1
pandas==1.1.2